+++ app/src/hooks/useCategories.ts
@@ -7,7 +7,7 @@ import { Category } from "@/schemas";
 export const useCategories = (query?: string) => {
     return useQuery({
         queryKey: ["categories", query],
-        queryFn: () => api.get<Category[]>(`/category/`, { params: { query } }),
+        queryFn: () => api.get<Category[]>("/category/", { params: { query } }),
     });
 };
