+++ app/src/routes/_mainLayout.tsx
@@ -1,28 +1,7 @@
-import { createFileRoute, Outlet } from "@tanstack/react-router";
-import Footer from "@/components/layout/footer";
-import ButtonNav from "@/components/layout/bottom-navbar";
-import { ChatBubble } from "@/components/store/chat-widget/ChatBubble";
-import StoreNavbar from "@/components/layout/shop-header";
-import FloatingActionDeck from "@/components/store/collections/floating-deck";
+import { createFileRoute } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/_mainLayout")({
-    component: MainLayoutComponent,
     errorComponent: ({ error }) => {
         throw error;
     },
 });
-
-function MainLayoutComponent() {
-    return (
-        <div className="flex flex-col min-h-screen">
-            <StoreNavbar />
-            <main className="flex-1 flex flex-col pt-[var(--nav-height)] pb-8">
-                <Outlet />
-            </main>
-            <Footer />
-            <ChatBubble />
-            <FloatingActionDeck />
-            <ButtonNav />
-        </div>
-    );
-}
