+++ backend/app/api/routes/product.py
@@ -49,7 +49,7 @@ async def get_recommendations(
 
 
 @router.get("/feed")
-@cacheable(key_prefix="products:list", tags=["products"], browser_ttl=60, cdn_ttl=600, cdn_swr=60)
+@cacheable(key_prefix="products:list", tags=["products"], browser_ttl=30, cdn_ttl=600, cdn_swr=60)
 async def feed(
     request: Request, srv: ProductDep, search: str = "", sort: str = "id:desc",
     cat_ids: str = Query(default=""), collections: str = Query(default=""),
@@ -95,7 +95,7 @@ async def search(
 
 @router.get("/{slug}")
 async def read(request: Request, slug: str, srv: ProductDep) -> ProductLite:
-    set_public_cache(request, browser_ttl=120, edge_ttl=86400, swr=600)
+    set_public_cache(request, browser_ttl=60, edge_ttl=86400, swr=600)
     cache_key: str = f"product:{slug}"
     cached = await srv.redis.get(cache_key)
     if cached:
