+++ backend/app/services/cache.py
@@ -257,7 +257,7 @@ def cacheable(
                 raw_key = f"{key_prefix}:{request.url.path}?{request.url.query}"
 
             if cdn_ttl is not None:
-                set_public_cache(request, browser_ttl=browser_ttl or 30, edge_ttl=cdn_ttl, swr=cdn_swr)
+                set_public_cache(request, browser_ttl=browser_ttl or 0, edge_ttl=cdn_ttl, swr=cdn_swr)
 
             cached_data = await cache.get(raw_key)
             if cached_data is not None:
